pass down only the loader label
