+++ frontend/src/components/GridWrapper/index.tsx
@@ -1,6 +1,5 @@
 import React, { ReactNode, useEffect, useState } from 'react';
 import classnames from 'classnames';
-import { Mode as GlobalMode, State } from 'store';
 import Grid, { Props as GridProps } from 'components/Grid';
 import Loader from 'components/Loader';
 import './styles.scss';
@@ -8,40 +7,18 @@ import './styles.scss';
 export type Props = {
   gridProps: GridProps;
   children: ReactNode;
-  apiFailed: State['apiFailed'];
-  fetchAbortController: State['fetchAbortController'];
-  mode: GlobalMode;
+  loaderLabel: string | null;
 };
 
 export default function GridWrapper({
   gridProps,
   children,
-  apiFailed,
-  fetchAbortController,
-  mode,
+  loaderLabel,
 }: Props) {
-  const showOverlay = !!fetchAbortController || !!apiFailed;
-  const [loaderLabel, setLoaderLabel] = useState('');
-
-  const getLoaderLabel = (): string | null => {
-    if (fetchAbortController) {
-      return mode === GlobalMode.EnterQuestions
-        ? 'Solving...'
-        : 'Generating...';
-    }
-    if (apiFailed === GlobalMode.Draw) {
-      return 'We were unable to generate the questions :(';
-    }
-    if (apiFailed === GlobalMode.EnterQuestions) {
-      return 'We couldn’t solve the crossword :(';
-    }
-
-    return null;
-  };
-
+  const [currentLoaderLabel, setCurrentLoaderLabel] = useState('');
   useEffect(() => {
-    setLoaderLabel(getLoaderLabel() ?? loaderLabel);
-  }, [fetchAbortController, apiFailed]);
+    setCurrentLoaderLabel(loaderLabel ?? currentLoaderLabel);
+  }, [loaderLabel]);
 
   return (
     <div className='grid-wrapper'>
@@ -49,10 +26,10 @@ export default function GridWrapper({
         <div
           className={classnames(
             'grid-wrapper__overlay',
-            !showOverlay && 'grid-wrapper__overlay_hidden',
+            !loaderLabel && 'grid-wrapper__overlay_hidden',
           )}
         >
-          <Loader label={loaderLabel} display={showOverlay} />
+          <Loader label={currentLoaderLabel} display={!!loaderLabel} />
         </div>
         <Grid {...gridProps} />
       </div>
